Single .js file between android and ios.
